Updated to 0.3

+++ m3/drppc/mmap.h
@@ -8,7 +8,7 @@
 #define INCLUDED_MMAP_H
 
 #include "drppc.h"
-#include "target.h"
+#include TARGET_CPU_HEADER
 
 #ifdef TARGET_IMPLEMENTS_MMAP_HANDLING
 
@@ -49,11 +49,11 @@ extern DRPPC_REGION *	MMap_FindWrite8Region(UINT32);
 extern DRPPC_REGION *	MMap_FindWrite16Region(UINT32);
 extern DRPPC_REGION *	MMap_FindWrite32Region(UINT32);
 
-extern UINT8			MMap_GenericRead8(UINT32);
-extern UINT16			MMap_GenericRead16(UINT32);
+extern UINT32			MMap_GenericRead8(UINT32);
+extern UINT32			MMap_GenericRead16(UINT32);
 extern UINT32			MMap_GenericRead32(UINT32);
-extern void				MMap_GenericWrite8(UINT32, UINT8);
-extern void				MMap_GenericWrite16(UINT32, UINT16);
+extern void				MMap_GenericWrite8(UINT32, UINT32);
+extern void				MMap_GenericWrite16(UINT32, UINT32);
 extern void				MMap_GenericWrite32(UINT32, UINT32);
 
 #endif // TARGET_IMPLEMENTS_MMAP_HANDLING
